вынос DrawableGameObject
