+++ packages/client/src/app/pages/game/gameEngine/store/gameState.ts
@@ -1,6 +1,6 @@
 import PlayerShip from '../objects/ships/playerShip';
-import params from '../parameters/gameParameters';
-import { NEXT_SHIP_DELAY, TEnemyType } from '../types/commonTypes';
+import params, { NEXT_SHIP_DELAY } from '../parameters/gameParameters';
+import { TEnemyType } from '../types/commonTypes';
 import GameLevels, { GameLevelList } from '../parameters/gameLevels';
 import Trajectory from '../objects/trajectory';
 import { GlobalGameState } from './objectState';
